pyproject.toml
[project]
name = "intern"
version = "0.1.0"
description = "A tool for helping your daily workflow"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "duckduckgo-search>=8.0.0",
    "langchain-anthropic>=0.3.10",
    "langchain-community>=0.3.21",
    "langgraph>=0.3.27",
    "python-dotenv>=1.1.0",
    "ruff>=0.11.4",
    "streamlit==1.44.0",
]
